fix everything, Makefile

+++ ft_memmove.c
@@ -10,7 +10,6 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <printf.h>
 #include "libft.h"
 
 /*
@@ -28,7 +27,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 {
 	unsigned char	*dest_ptr;
 	unsigned char	*src_cpy;
-	int				i;
+	size_t			i;
 
 	dest_ptr = (unsigned char *)dst;
 	src_cpy = (unsigned char *)src;
